refactor: 훅 이름 변경된 이름으로 수정
